+++ addons/YarnSpinner-Godot/Runtime/Views/OptionsListView.cs
@@ -1,6 +1,7 @@
 #nullable disable
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Godot;
 
@@ -76,7 +77,7 @@ namespace YarnSpinnerGodot
             {
                 // prevent option views from being pressed by the same input that advanced the dialogue
                 // by waiting a frame
-                var mainTree = (SceneTree)Engine.GetMainLoop();
+                var mainTree = (SceneTree) Engine.GetMainLoop();
                 await mainTree.ToSignal(mainTree, SceneTree.SignalName.ProcessFrame);
                 viewControl.Visible = false;
                 // Hide all existing option views
@@ -152,6 +153,7 @@ namespace YarnSpinnerGodot
                 // Note the delegate to call when an option is selected
                 OnOptionSelected = onOptionSelected;
 
+                viewControl.Visible = true;
                 // Fade it all in
                 await Effects.FadeAlpha(viewControl, 0, 1, fadeTime);
 
@@ -193,8 +195,8 @@ namespace YarnSpinnerGodot
                         OnOptionSelected(selectedOption.DialogueOptionID);
                     }
                 }
-
-                optionViews[0].GrabFocus();
+                // If the user is hiding unavailable options, select the first visible one.
+                optionViews.First(view => view.Visible).GrabFocus();
             }
             catch (Exception e)
             {